# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3001
info:
  contact: {}
  description: API untuk sistem pelaporan prestasi mahasiswa dengan dukungan multi-role
    (Admin, Dosen Wali, Mahasiswa) dan field prestasi yang fleksibel
  title: Sistem Pelaporan Prestasi Mahasiswa API
  version: "1.0"
paths:
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Mengambil daftar notifikasi user dengan pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get notifications
      tags:
      - Notifications
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Menandai notifikasi sebagai sudah dibaca berdasarkan ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/read-all:
    put:
      consumes:
      - application/json
      description: Menandai semua notifikasi user sebagai sudah dibaca
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Mark all notifications as read
      tags:
      - Notifications
  /notifications/unread-count:
    get:
      consumes:
      - application/json
      description: Mengambil jumlah notifikasi yang belum dibaca untuk user yang sedang
        login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get unread notification count
      tags:
      - Notifications
  /reports/lecturer:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi mahasiswa bimbingan untuk lecturer yang
        sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current lecturer report
      tags:
      - Reports
  /reports/lecturer/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi mahasiswa bimbingan untuk lecturer berdasarkan
        ID
      parameters:
      - description: Lecturer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get lecturer report by ID
      tags:
      - Reports
  /reports/statistics:
    get:
      consumes:
      - application/json
      description: 'Mengambil statistik achievement berdasarkan role (Mahasiswa: milik
        sendiri, Dosen Wali: mahasiswa bimbingan, Admin: semua)'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get achievement statistics
      tags:
      - Reports
  /reports/student:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi student yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current student report
      tags:
      - Reports
  /reports/student/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi student berdasarkan ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student report by ID
      tags:
      - Reports
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"